+++ craco.config.js
@@ -15,6 +15,7 @@ module.exports = {
       "@assets": path.resolve(__dirname, "src/assets"),
       "@utils": path.resolve(__dirname, "src/utils"),
       "@enums": path.resolve(__dirname, "src/enums"),
+      "@providers": path.resolve(__dirname, "src/providers"),
     },
   },
 };
