+++ src/components/Toast/styles.scss
@@ -1,6 +1,4 @@
 .toast {
-  right: 0;
-  margin-right: 10px;
   position: relative;
   align-items: center;
   gap: 10px;
@@ -10,6 +8,8 @@
   max-width: 220px;
   width: 100%;
   overflow: hidden;
+  z-index: 1000;
+  background-color: var(--bg-white-1);
 
   .toast-content {
     flex: 1;
@@ -45,12 +45,6 @@
   &--show {
     display: flex;
   }
-
-  &--hide {
-    display: none;
-  }
-
-  animation: slide linear 1.5s;
 }
 
 .progress-error::before {
@@ -59,12 +53,6 @@
 
 .progress-success::before {
   background-color: var(--bg-green-1);
-}
-
-@keyframes slide {
-  0% {
-    left: 100%;
-  }
 }
 
 @keyframes progress {
