--- src/contexts/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type { IToastContextValue } from "./IToastContextValue";
