+++ src/providers/toast-provider/interfaces/IToast.ts
@@ -1,0 +1,4 @@
+export interface IToast {
+  id: number;
+  message: string;
+}
