+++ src/providers/toast-provider/interfaces/IToastProviderProps.ts
@@ -1,0 +1,5 @@
+import type React from "react";
+
+export interface IToastProviderProps {
+  children: React.ReactElement;
+}
