+++ src/providers/toast-provider/styles.scss
@@ -1,0 +1,8 @@
+.toasts {
+  position: fixed;
+  right: 0;
+  top: 30px;
+  padding: 0 40px;
+  max-width: 220px;
+  width: 100%;
+}
