+++ tsconfig.json
@@ -28,7 +28,8 @@
       "@server": ["./server/index"],
       "@interfaces": ["./interfaces/index"],
       "@assets": ["./assets/index"],
-      "@enums": ["./enums/index"]
+      "@enums": ["./enums/index"],
+      "@providers": ["./providers/index"]
     }
   },
   "include": ["src"]
